requirements_linux.txt
aiofiles==24.1.0
aiogram==3.17.0
aiogram_calendar==0.6.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
blinker==1.9.0
branca==0.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ConfigSpace==1.2.1
cryptography==44.0.0
dadata==24.10.0
Deprecated==1.2.18
dill==0.3.9
et_xmlfile==2.0.0
exceptiongroup==1.2.2
Flask==3.1.0
folium==0.19.4
frozenlist==1.5.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jpype1==1.5.1
magic-filter==1.0.12
MarkupSafe==3.0.2
more-itertools==10.6.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
polyline==2.0.2
propcache==0.2.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyvroom==1.14.0
redis==5.2.1
requests==2.32.3
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
unified-planning==1.2.0
up-jsprit==1.0.5
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
xyzservices==2024.9.0
yandexcloud==0.337.0
yarl==1.18.3
